refactor test server in to common code for agent (#31869)
* wip


* lint
